+++ Frames/Contracts/Repositories/IPaymentsRepository.cs
@@ -3,9 +3,9 @@ namespace Frames.Contracts.Repositories;
 public interface IPaymentsRepository
 {
     Task<List<Payments>> GetPayments(int month, int year);
-    Task<Payments?> GetPayment(int paymentId);
+    Task<Payments?> GetPayment(long paymentId);
     Task<Payments?> GetPayment(DateOnly date);
     void CreatePayment(Payments payment);
     void UpdatePayment(Payments payment);
-    void DeletePayment(int paymentId);
+    void DeletePayment(long paymentId);
 }
