+++ Frames/Entities/Paid.cs
@@ -6,8 +6,8 @@ public class Paid : EntityBase
 {
     public DateOnly Date { get; set; }
     public decimal Amount { get; set; }
-    public int PaymentId { get; set; }
-    public int BillingId { get; set; }
+    public long PaymentId { get; set; }
+    public long BillingId { get; set; }
 
     [ForeignKey(nameof(BillingId))]
     public Billing? Billing { get; set; }
